refactor(entities): isolated entities visual instead of putting it in entities
