--- Assets/Code/Scripts/Entities/EntityConfigurationSO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9bc562c7cdab31240b6e99c3650435c8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
