--- Assets/ScriptableObjects/Entities.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: acf3616d3c7c3ca4fa8b64a6535cc07a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
