--- Assets/ScriptableObjects/Entities/HeavyWarriorEntityConfiguration.asset
@@ -1,17 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 9bc562c7cdab31240b6e99c3650435c8, type: 3}
-  m_Name: HeavyWarriorEntityConfiguration
-  m_EditorClassIdentifier: 
-  <Name>k__BackingField: Warrior
-  <Icon>k__BackingField: {fileID: 0}
-  <DefaultAnimationState>k__BackingField: Idle
