--- Assets/ScriptableObjects/Entities/HeavyWarriorEntityConfiguration.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cadcc449204f6ac44931b3f48473a642
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
